titanic-backend/requirements.txt
fastapi
uvicorn
pydantic>=2.0
langchain>=0.2.0,<0.3.0
langchain-groq
langchain-experimental
langchain-community
langchain-chroma
langchain-huggingface
langchain-core
langsmith
chromadb
pandas
matplotlib
seaborn
sentence-transformers
python-dotenv
langchain-text-splitters
--extra-index-url https://download.pytorch.org/whl/cpu
torch
